src/forecast/requirements.txt
pandas==1.1.5
numpy==1.19.5
scikit-learn==0.24.1
scipy==1.5.4
joblib==1.0.0
pytest==6.2.2
statsmodels==0.12.1
coverage
flake8
matplotlib==3.3.4
cassandra-driver==3.25.0